requirements.txt
django
django-stdimage
django-crispy-forms

pylint
pylint-django
flake8

ipython